requirements.txt
flask==3.0.0
chromadb==0.4.15
pymongo==4.5.0
flask-pymongo==2.3.0
python-dotenv==1.0.0
google-generativeai>=0.3.2
python-dateutil==2.8.2


